Add code formatting check to Travis CI build
Check compliance with the code style standard established by Arduino's examples_formatter.conf Artistic Style configuration file:
https://github.com/arduino/Arduino/blob/master/build/shared/examples_formatter.conf
If the code doesn not comply with the standard, a diff between the submitted code and the auto-formatted code is shown in the log.

I did some reorganization of .travis.yml so that the formatting check was done in a single dedicated Travis CI job, rather than redundantly in every one of the jobs.