Manually improve some formatting left ugly by the auto format
All the breaks were originally aligned. On some lines, this was accomplished by adding spaces before the closing parentheses, but not all lines have one. Because the auto format removes padding inside parentheses, this resulted in the spaces being removed from the lines with the parentheses, but not on the other lines. I chose to make the most minimal fix, but the alternative could be to add spaces outside the parentheses to recover alignment of all the breaks, which is fully compliant with the formatting check.