Use arduino-cli for Travis CI build
Use arduino-cli instead of the Arduino IDE for the Travis CI build.

arduino-cli does not result in the many "SocketListener" warnings that end up interspersed with the compilation output when using the Arduino IDE with Travis CI. These warnings make it more likely that Travis CI's maximum job length will be exceeded, resulting in an errored job. The warnings also make the build log hard to read.

Due to a smaller download size, using arduino-cli results in a shorter CI build duration.