Register a callback function with ArduinoCloudThing (#93)
This allows class ArduinoCloudThing to obtain a 'epoch' timestamp which is used during synchronisation and which has been implemented so far via an 'extern' dependency to RTCZero. Since this kind of hidden dependency (basically a global object across 2 separate Arduino libraries) is very brittle the depdency is made clear by explicitly registering a function to provide the required timestamp.