Improve board-specific configuration system in compile-examples CI workflow (#124)
Use the matrix.include key to customize the compile-examples CI workflow's matrix jobs with board type-specific configurations.

This GitHub Actions feature allows defining additional properties for matrix jobs that match the filter.

The previous approach for doing this was to use generated environment variable names, which worked but seemed likely to be confusing for anyone trying to decipher the workflow. I think the new approach is a little easier to understand and less hacky.

I had hoped this would allow me to do away with the board.type system, but supplying lists of values to the filter keys isn't supported, you can't reference the env key from matrix, and YAML anchors aren't supported. So the board.type system remains the only way I can see to avoid redefining the configuration options for each board of the same type.