Remove unnecessary entries from spell check ignore list
These entries were previously necessary to ignore misspelled words in the external resources contained in the repository because the spell check action didn't have the ability to exclude those paths. Since then, the "skip-paths" input was added to the action, so these entries only increase the chances of false negatives from the spell check CI workflow.