Refresh spell check CI workflow (#134)
* Fix YAML formatting of spell check CI workflow

An unintentional extra indent made the YAML not a valid GitHub Actions workflow.

* Fix misspelled word in comment

This was causing the spell check CI workflow to fail.

* Configure spell check CI workflow to ignore name causing a false positive

* Configure spell check CI workflow to skip folder containing externally-maintained resources

A misspelled word in the tinycbor library was causing the spell check CI workflow to fail. Since this code is not maintained in this repository, it is best to just exclude that folder from the misspelled words check.

* Remove unnecessary entries from spell check ignore list

These entries were previously necessary to ignore misspelled words in the external resources contained in the repository because the spell check action didn't have the ability to exclude those paths. Since then, the "skip-paths" input was added to the action, so these entries only increase the chances of false negatives from the spell check CI workflow.