Use sketch compilation CI actions from their new dedicated repositories
The "smoke test" sketch compilation and reporting actions have been moved to dedicated repositories. The actions hosted
at the old `arduino/actions/libraries/*` location will no longer be maintained.

Since the time the actions were migrated to the dedicated repositories, a breaking change was made to the default value
of the `sketches-report-path` input, which required a change to the values of the `name` and `path` inputs of the
	`actions/upload-artifact` action.