Specify sketches report input values for sketch compilation CI actions
The reliance on the `actions/upload-artifacts` action means that these values will always be hardcoded into the
workflow. The result is that if the default values of the inputs are changed, it breaks the workflow, even if the
changes are synchronized. So it's safest to not rely on the defaults.