Use modern API of `arduino/compile-sketches` CI action
Since the time the "Compile Examples" workflow was written, some improvements have been made to the API of the
`arduino/compile-sketches` CI action. The previous API is still supported, but is deprecated and warnings are displayed
in the workflow run log about this, which could result in confusion for contributors.