Pin Arduino CLI version used in CI to 0.13.0 (#222)
When the version is not pinnec, the latest release is used (currently 0.14.0). This version has an incompatibility with
the combination of ArduinoIoTCloud and ESP8266. The easy workaround is to use the previous Arduino CLI version to get
the CI system working again.

Once the real fix for the issue is found, this commit should be reverted so the CI system can benefit from the ongoing
development work on Arduino CLI.