Use official GitHub Actions action for spell checking
The `arduino/actions/libraries/spell-check` GitHub Actions action is now deprecated in favor of the official action from
codespell.

The 2.0.0 release of codespell added support for a .codespellrc configuration file. Use of this file to configure
codespell allows the spell check to be run locally with the identical configuration to that used by the CI system.