Track the number of unsuccessful request-last-values. (#249)
In the current implementation the number the board would continously attempt to obtain the last values from the cloud. If, however, those values never arrive - then the board would never move on and become operational. It's therefore better to count the number of failed attempts and reset the connection to allow a clean re-connection from the ground up.