Correct library dependency name in metadata
The `depends` field of the library.properties metadata file can be used to provide a list of library dependencies that should be installed in addition during a Library Manager installation. The library names must match exactly to the `name` value in the dependency's library.properties metadata file, which is the unique identifier used by Library Manager.

The name of the "Adafruit SleepyDog Library" is just like that, but an incorrect name "Adafruit_SleepyDog" was used instead. This resulted in the following:

- Classic Arduino IDE: the incorrect name was ignored and the dependency not installed
- Arduino IDE 2.x: installation of the ArduinoIoTCloud library fails silently
- Arduino CLI: installation of the ArduinoIoTCloud library fails with "Error installing ArduinoIoTCloud: No valid dependencies solution found: dependency 'Adafruit_SleepyDog' is not available"