Add comment to "Compile Examples" workflow re: esp8266:esp8266 version
A specific version of the "esp8266" boards platform is installed on the Arduino Cloud servers.

Since that is the primary environment the sketches using this library are compiled in, it is necessary for the "Compile
Examples" GitHub Actions workflow to use the same version of the platform when running the test compilations of the
library examples in order to ensure the results from the CI system will match those from real usage.

This requirement is not obvious. Since the workflow does not pin the other platform versions (because Arduino Cloud
typically uses the latest version), the pinning of this platform specifically might be the source of confusion for the
maintainers. So it is worth documenting the reason.