Exclude BearSSL library files from spell check
The latest release of the codespell tool used for automated spell checking of the files of this project included an
expansion of the misspelled words dictionary.

This caused the spell checker to trigger on some content in the BearSSL library vendored in this project. These included
both false and true positives. Although the former can be fixed by ignoring the specific word and the latter by fixing
the typo, it doesn't really make sense to be checking these files which appear to be mostly copies of content maintained
externally.

For this reason, the best approach will be to exclude the files from the check entirely.