Fix spell check false positive by ignoring word
In the latest release of the codespell tool used for automated spell checking of the files of this project, the word
"bu" was added to the codespell misspelled words dictionary as a misspelling of "by".

This caused a false detection of a CBOR map label `bu` as a misspelling, resulting in a failing spell check result.

Since the occurrence of this word is correct and intended in this project, the false positive is resolved by configuring
codespell to ignore the problematic word.