Bump `codecov/codecov-action` dependency to v3
The "Unit Test" GitHub Actions workflow uses the "codecov/codecov-action" GitHub Actions action to upload coverage data
to Codecov.

Previously the workflow specified the "v1" version ref for the action dependency. That pinned the dependency to the
major version 1 series. Several major version bumps have been made in the action since that time so the use of the v1
ref caused the workflow to use a significantly outdated and deprecated version of the action.

Bumping the action to the latest "v3" ref will avoid the potential for the step to stop working entirely (the action
docs claim v1 was sunset 1.5 years ago) and allow the workflow to benefit from the enhancements and fixes from ongoing
development work in the action until such time as a breaking change in the action triggers them to make another major
version bump. At that time it will be necessary for the project maintainers to evaluate whether the breaking change
requires any modifications to the workflow before manually bumping the ref again (i.e.,
`uses: codecov/codecov-action@v3` -> `uses: codecov/codecov-action@v4`).