Revert "Fix regression: report size delta size on PR. (#429)"
This partially reverts commit 836771a7a4a2d49758bf7e977f68fa2b124a5941.

The repository infrastructure uses the "arduino/report-size-deltas" action to generate a report of the changes in memory
usage of the example sketches that would result from merging a pull request.

There are two use patterns for the "arduino/report-size-deltas" action:

* Run from a workflow triggered by a `schedule` event.
* Run from the same workflow as the "arduino/compile-sketches" action when triggered by a `pull_request` event.

The latter use pattern is only suitable for private repositories. The reason is that, when a pull request is submitted
from a fork, the permissions of the access token used by the "arduino/report-size-deltas" action are downgraded to
read-only in workflows triggered by a `pull_request` event, which means it is unable to make the report comment.

The reverted commit migrated the size deltas report infrastructure from the first use pattern to the second, which
caused the "Compile Examples" workflow runs to fail on pull requests submitted from forks:

Error: HTTPError: HTTP Error 403: Forbidden

For this reason, the infrastructure must be reverted back to using the first use pattern.